Move base type declarations to matcher and use-location.
