+++ types/matcher.d.ts
@@ -1,4 +1,24 @@
-import { MatcherFn } from "./index";
+/*
+ * Match and params
+ */
+
+import { Path } from "./use-location";
+
+export interface DefaultParams {
+  [paramName: string]: string;
+}
+export type Params<T extends DefaultParams = DefaultParams> = T;
+
+export type MatchWithParams<T extends DefaultParams = DefaultParams> = [
+  true,
+  Params<T>
+];
+export type NoMatch = [false, null];
+export type Match<T extends DefaultParams = DefaultParams> =
+  | MatchWithParams<T>
+  | NoMatch;
+
+export type MatcherFn = (pattern: Path, path: Path) => Match;
 
 export default function makeMatcher(
   makeRegexpFn?: (
